+++ config/blocks.php
@@ -64,14 +64,14 @@ return [
                     'source' => A17\Twill\Services\Blocks\Block::SOURCE_TWILL,
                 ],
                 [
-                    'path' => resource_path('views/twill/admin/blocks'),
+                    'path' => resource_path('views/twill/blocks'),
                     'source' => A17\Twill\Services\Blocks\Block::SOURCE_APP,
                 ],
             ],
 
             'repeaters' => [
                 [
-                    'path' => resource_path('views/twill/admin/repeaters'),
+                    'path' => resource_path('views/twill/repeaters'),
                     'source' => A17\Twill\Services\Blocks\Block::SOURCE_APP,
                 ],
                 [
@@ -82,16 +82,16 @@ return [
 
             'icons' => [
                 base_path('vendor/area17/twill/frontend/icons'),
-                resource_path('views/twill/admin/icons'),
+                resource_path('views/twill/icons'),
             ],
         ],
 
         'destination' => [
             'make_dir' => true,
 
-            'blocks' => resource_path('views/twill/admin/blocks'),
+            'blocks' => resource_path('views/twill/blocks'),
 
-            'repeaters' => resource_path('views/twill/admin/repeaters'),
+            'repeaters' => resource_path('views/twill/repeaters'),
         ],
     ],
 ];