+++ tests/integration/CapsuleAutoloadTest.php
@@ -31,11 +31,11 @@ class CapsuleAutoloadTest extends TestCase
         '{$stubs}/capsules/posts/Models/PostRevision.php' =>
             '{$app}/Twill/Capsules/Posts/Models/PostRevision.php',
 
-        '{$stubs}/capsules/posts/resources/views/twill/admin/form.blade.php' =>
-            '{$app}/Twill/Capsules/Posts/resources/views/twill/admin/form.blade.php',
+        '{$stubs}/capsules/posts/resources/views/twill/form.blade.php' =>
+            '{$app}/Twill/Capsules/Posts/resources/views/twill/form.blade.php',
 
-        '{$stubs}/capsules/posts/resources/views/twill/admin/create.blade.php' =>
-            '{$app}/Twill/Capsules/Posts/resources/views/twill/admin/create.blade.php',
+        '{$stubs}/capsules/posts/resources/views/twill/create.blade.php' =>
+            '{$app}/Twill/Capsules/Posts/resources/views/twill/create.blade.php',
 
         '{$stubs}/capsules/posts/routes/twill.php' =>
             '{$app}/Twill/Capsules/Posts/routes/twill.php',